requirements.txt
aiogram
ccxt
python-dotenv
